Move beOwnedBy test to common file
